'feat(coc)!: when pum is visible, <cr> to select it
Before, the sequence of typing is: type a little, <tab> to navigate, and words are showing. No <cr> is required.
Now, I'm changing the work flow to: type a little, <tab> not really required to find the word, <cr> to use it. This way is going to better integrate with snippets.